+++ NEWS.md
@@ -1,3 +1,7 @@
+# cookiemonster 0.0.4
+
+* adds support for cookies from rvest::read_html_live sessions
+
 # cookiemonster 0.0.3
 
 * fixes issue with "HTTPonly" cookies